requirements.txt
pyrogram>=2.0.106
tgcrypto>=1.2.5
python-dotenv>=1.0.1
motor>=3.6.0
aiofiles>=24.1.0
